+++ packages/grant-permission-dialog/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Localize available action names
+
 ## 0.1.0
 
 - Initial release